add the 'compact' formatter
was removed when implementing protofire/solhint/issues/385 , I probably
missed it since it wasn't listed in the docs or command line options.
It's likely someone else will show up asking for some other eslint
formatter to be added back.

This adds it back from eslint @ 8.42.0.

I also ran our linter on it, so it's compliant with this codebase's
style