fix: recognize usages as values
Fixes #19 #20

I'm not fully convinced this is going to be robust. Any non-import node
with an identifier with the same name would mark the name as used, which
could cause an unused import to not be reported if it's shadowed.

Some listeners only registered the name of a child Identifier, so
they're made redundant now.