explicit-type tests: only ever enforce explicit types
I didn't really research if someone asked for a rule to require implicit
types, but I don't think it's very idiomatic for a rule called
'explicit-types' to check that implicit types are used.

it's sad that this'll break compatibility with protofire/solhint, but it
enables me to remove a lot of code, which happens to be buggy. For
example, this test:

it('should throw only ONE error with a wrong configuration rule and two variables', () => {
    const code = contractWith('uint256 public constant SNAKE_CASE = 1; uint256 public foo; ')

    const report = linter.processStr(code, {
    rules: { 'explicit-types': ['error', 'implicito'] },
    })

    assertErrorCount(report, 1)
})

wouldnt pass with the current implementation. And there are some other
issues that'll get their own commit.