reimplement explicit-types in a more robust way
the previous implementation also visited the AST manually, descending
into all the sub-objects of a few AST nodes. Although the algorithm
started from nodes which were unlikely to have a big tree behind them
(the most problematic perhaps being ExpressionStatement), it was not at
all elegant and liable to slow down the linter.

This one, aside from dropping the possibility of requiring implicit
types, visits nodes that can use a type explicitly, albeit for cases
like 'uint256[] public arr = [uint(1),2,3];' I got two calls to the
FunctionCall callback, which I didn't expect and caused the same line to
be reported twice. This is perhaps why Diego chose to walk the AST
manually. I instead added a Set to not report on the same node twice,
but I should figure out if something is wrong with either the AST
visitor or what I expect from it.