ensure docs are up-to-date
It's too easy to forget to run `npm run docs` when modifying/creating a
rule and then have another unrelated commit two weeks later include the
changes to the compiled docs, which pollutes git history. This should
check we update the docs before merging a PR

I had to add a fetch-depth: 0 to the e2e job because otherwise the
GitHelper in script/generate-rule-docs.js couldn't get when the file for
a particular rule was added, and therefore modified all {rule-name}.md
files to say 'this rule was added in the last version'.