modify loadConfig so it can load the config file from a directory
previously it only took an optional filename, which was the value of the
--config flag, and if it wasn't present then it searched the current working
directory with cosmiconfig.

This changes enables us to do the same with arbitrary directories, to
support reading config files from subdirectories as described in #84

I also added the listConfigsForPaths function, which should figure out
the list of directories where configs should be read for a given file.
See the tests for details.