Actually remove solhint:default ruleset
in #103 I forgot to actually remove the files defining the config. That
would've been avoided by having some code bro review the PR.

tests pass without the changes in this pr because I had already removed
the `solhit:default` from getSolhintCoreConfig. So this is more of a
'remove dead code' thing