new rule definition-name-capwords
with the functionality of contract-name-camelcase and event-name-camelcase, and extra testing.

It reports errors on the same cases, but with better messages, docs and
testing

Also, the formal name for the pattern is (at least in the style guide)
CapWords, for capitalized camelcase. So I changed it to that.

initially I divided contract-name-camelcase into:
- struct-name-capwords
- enum-name-capwords
- event-name-capwords (replacing event-name-camelcase)
- contract-name-capwords

but was not sold on this, I don't know why anyone would
want one rule but not the other 3