make func-param-name-mixedcase a singular case of var-name-mixedcase
at the AST and conceptual level, function parameters are Identifiers,
and the existing var-name-mixedcase rule already covered the cases where
func-param-name-mixedcase would error (note I didn't modify behaviour on
the former). So it's of no value to have the latter.

I just moved the testcases from func-param-name-mixedcase to work as
regressions and deleted the rule.