extra tests for no-unused-vars
currently no-unused-vars doesn't report errors on unused state variables.
wether we should or not, that's another question, but I should not
modify that behaviour in the process of merging it with no-unused-import

current bug in no-unused-vars:

this is currently mitigated by the fact that `solc` warns on shadowing,
but it's nevertheless a bug with the rule. To fix this, the
VarUsageScope should be attached to the block scope of the function, and
new ones created for other block scopes. We'll probably have to figure
out to skip some defined block scopes (contract definitions? idk they
have curly brackets)

And we should also implement the logic to keep looking for a funcScope
when finding one and realising the variable we want to track isn't
there. That'll be crucial to processing imports correctly.